+++ source/packages/@aws-accelerator/accelerator/lib/config-validator.ts
@@ -31,7 +31,8 @@ import {
   SecurityConfigValidator,
   ReplacementsConfigValidator,
 } from '@aws-accelerator/config';
-import { createLogger, setExternalManagementAccountCredentials } from '@aws-accelerator/utils';
+import { createLogger, getExternalManagementAccountCredentials } from '@aws-accelerator/utils';
+import { AwsCredentialIdentity } from '@aws-sdk/types';
 
 const logger = createLogger(['config-validator']);
 const configDirPath = process.argv[2];
@@ -124,8 +125,9 @@ async function validateConfig(props: {
   stage: string | undefined;
 }) {
   let orgsEnabled: boolean;
+  let managementAccountCredentials: AwsCredentialIdentity | undefined;
   try {
-    await setExternalManagementAccountCredentials(props.partition, homeRegion);
+    managementAccountCredentials = await getExternalManagementAccountCredentials(props.partition, homeRegion);
     orgsEnabled = OrganizationConfig.loadRawOrganizationsConfig(configDirPath).enable;
   } catch (e) {
     logger.error('Failure in validationConfig', e);
@@ -134,7 +136,13 @@ async function validateConfig(props: {
   let accountsConfig: AccountsConfig | undefined = undefined;
   try {
     accountsConfig = AccountsConfig.load(configDirPath);
-    await accountsConfig.loadAccountIds(props.partition, props.enableSingleAccountMode, orgsEnabled!, accountsConfig);
+    await accountsConfig.loadAccountIds(
+      props.partition,
+      props.enableSingleAccountMode,
+      orgsEnabled!,
+      accountsConfig,
+      managementAccountCredentials,
+    );
   } catch (e) {
     initErrors.push({ file: AccountsConfig.FILENAME, message: e });
   }